% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lsos.R
\name{.ls.objects}
\alias{.ls.objects}
\title{Title https://stackoverflow.com/a/47001817}
\usage{
.ls.objects(pos = 1, pattern, order.by, decreasing = FALSE, head = FALSE,
  n = 5)
}
\arguments{
\item{n}{}
}
\description{
Title https://stackoverflow.com/a/47001817
}
