% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/splitAt.R
\name{splitAt}
\alias{splitAt}
\title{https://stackoverflow.com/questions/16357962/r-split-numeric-vector-at-position}
\usage{
splitAt(x, pos)
}
\arguments{
\item{pos}{}
}
\description{
https://stackoverflow.com/questions/16357962/r-split-numeric-vector-at-position
}
